requirements/local.txt
Django==4.2.4
djangorestframework==3.14.0
django-redis==5.3.0
pyjwt